Fix device scanning in DevicePickerFragment
In some conditions, the BT devices scan may be blocked with the phone
displaying "no devices detected".

This behavior is reproduced on Galaxy Nexus reference phone.

This patch will fix the behavior in most of the scenarios by doing a
device scan by default.
However, the device scan will still be blocked upon screen rotation as
this is the expected behavior.

Change-Id: Ic999355db93eeeca97435da9d894c13e4bd97ce8
Orig-Change-Id: Iaef4d8857b006de1e68854d23e0810a433727f4d
Signed-off-by: Patrick Combes <patrick.combes@intel.com>
Reviewed-on: http://android.intel.com:8080/119743
Reviewed-by: Boniface, XavierX <xavierx.boniface@intel.com>
Reviewed-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Reviewed-by: Aquilone, RaffaeleX <raffaelex.aquilone@intel.com>
Reviewed-by: Gros, Julien <julien.gros@intel.com>
Tested-by: Lestienne, ChristopheX <christophex.lestienne@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: Jian Luo <jian.luo@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 124172